Move useDriverStore into driver folder
